my_hit_backend/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.2
django-crispy-forms==1.8.1
django-multiselectfield==0.1.11
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.8
mysqlclient==1.3.12
ndg-httpsclient==0.5.1
Pillow==7.0.0
pusher==2.1.4
pyasn1==0.4.8
pycparser==2.19
pyfcm==1.4.7
PyNaCl==1.3.0
pyOpenSSL==19.1.0
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.34
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
webencodings==0.5.1
xhtml2pdf==0.2.4
